requirements.txt
paho-mqtt
influxdb
influxdb-client[ciso]
marshmallow